refactor: centralize purchase logic
