+++ kiosk-backend/routes/admin/buy_for_user.js
@@ -1,5 +1,6 @@
 import express from 'express';
 import supabase from '../../utils/supabase.js';
+import purchaseProduct from '../../utils/purchaseProduct.js';
 const router = express.Router();
 
 router.post('/', async (req, res) => {
@@ -20,31 +21,15 @@ router.post('/', async (req, res) => {
     .single();
 
   if (!user || !product || product.stock < quantity) {
-    return res.status(400).json({ error: 'Nicht genügend Bestand oder Nutzer nicht gefunden' });
+    return res
+      .status(400)
+      .json({ error: 'Nicht genügend Bestand oder Nutzer nicht gefunden' });
   }
 
-  const total = quantity * product.price;
-  const newBalance = (user.balance || 0) - total;
+  const { error, success } = await purchaseProduct(user, product, quantity);
 
-  const { error: purchaseError } = await supabase.from('purchases').insert({
-    user_id: user.id,
-    user_name: user.name,
-    product_id,
-    product_name: product.name,
-    price: total,
-    quantity
-  });
-  const { error: balanceError } = await supabase
-    .from('users')
-    .update({ balance: newBalance })
-    .eq('id', user.id);
-  const { error: stockError } = await supabase
-    .from('products')
-    .update({ stock: product.stock - quantity })
-    .eq('id', product_id);
-
-  if (purchaseError || balanceError || stockError) {
-    return res.status(500).json({ error: 'Fehler beim Kaufvorgang' });
+  if (!success) {
+    return res.status(500).json({ error });
   }
 
   res.json({ success: true });
