+++ kiosk-backend/routes/buy.js
@@ -1,6 +1,7 @@
 import express from 'express';
 import supabase from '../utils/supabase.js';
 import getUserFromRequest from '../utils/getUser.js';
+import purchaseProduct from '../utils/purchaseProduct.js';
 const router = express.Router();
 
 router.post('/', async (req, res) => {
@@ -12,34 +13,24 @@ router.post('/', async (req, res) => {
     return res.status(400).json({ error: 'Ungültige Eingaben' });
   }
 
+  const { data: dbUser } = await supabase
+    .from('users')
+    .select('*')
+    .eq('id', user.id)
+    .single();
+  const { data: product } = await supabase
+    .from('products')
+    .select('*')
+    .eq('id', product_id)
+    .single();
 
-  const { data: dbUser } = await supabase.from('users').select('*').eq('id', user.id).single();
-  const { data: product } = await supabase.from('products').select('*').eq('id', product_id).single();
+  if (!product || product.stock < quantity)
+    return res.status(400).json({ error: 'Nicht genügend Bestand' });
 
-  if (!product || product.stock < quantity) return res.status(400).json({ error: 'Nicht genügend Bestand' });
+  const { error, success } = await purchaseProduct(dbUser, product, quantity);
 
-  const total = quantity * product.price;
-  const newBalance = dbUser.balance - total;
-
-  const [
-    { error: purchaseError },
-    { error: balanceError },
-    { error: stockError }
-  ] = await Promise.all([
-    supabase.from('purchases').insert({
-      user_id: dbUser.id,
-      user_name: dbUser.name,
-      product_id,
-      product_name: product.name,
-      price: total,
-      quantity
-    }),
-    supabase.from('users').update({ balance: newBalance }).eq('id', dbUser.id),
-    supabase.from('products').update({ stock: product.stock - quantity }).eq('id', product_id)
-  ]);
-
-  if (purchaseError || balanceError || stockError) {
-    return res.status(500).json({ error: 'Fehler beim Kaufvorgang' });
+  if (!success) {
+    return res.status(500).json({ error });
   }
 
   res.json({ success: true });
